LayoutTests/platform/chromium/fast/js/function-declaration-expected.txt

This test ensures we correctly parse (or not) various function declarations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS eval('function f(){return true;}') is undefined.
PASS eval('function f(){return true;};f')() is true
PASS eval('function(){return false;}')() threw exception SyntaxError: Unexpected token (.
PASS successfullyParsed is true

TEST COMPLETE
